Refactor dataset testing with language support
